# contracts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nebulosa GitHub Architecture Agent API",
    "version": "1.0.0",
    "description": "REST API for the codebase-aware software architecture agent."
  },
  "paths": {
...
    "/v1/tasks": {
      "post": {
        "summary": "Create architecture task",
        "operationId": "create_task",
        "tags": ["tasks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["repository_id", "task_type", "prompt"],
                "properties": {
                  "repository_id": { "type": "string", "format": "uuid" },
                  "task_type": {
                    "type": "string",
                    "enum": [
                      "architecture_report", "architecture_problems", "target_architecture",
                      "service_boundaries", "folder_structure", "database_design",
                      "api_design", "deployment_design", "roadmap", "repo_compare"
                    ]
                  },
                  "prompt": { "type": "string", "maxLength": 4096 },
                  "snapshot_id": { "type": "string", "format": "uuid", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "202": { "description": "Task accepted" },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/v1/tasks/{taskId}": {
      "get": {
        "summary": "Get task status",
        "operationId": "get_task",
        "tags": ["tasks"],
        "parameters": [
          { "name": "taskId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Task status" },
          "404": { "description": "Task not found" }
        }
      }
    },
    "/v1/tasks/{taskId}/report": {
      "get": {
        "summary": "Get task report",
        "operationId": "get_task_report",
        "tags": ["tasks"],
        "parameters": [
          { "name": "taskId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Architecture report" },
          "404": { "description": "Report not found" }
        }
      }
    },
...
  }
}